requirements.txt
pycryptodome==3.21.0
pytest==8.3.4